Set up ChartContext in each package and refactor ScrubberProvider to encapsulate all scrubbing logic
