+++ packages/mobile-visualization/src/chart/axis/XAxis.tsx
@@ -4,8 +4,8 @@ import type { ThemeVars } from '@coinbase/cds-common';
 import {
   getAxisTicksData,
   isCategoricalScale,
-  useChartContext,
 } from '@coinbase/cds-common/visualizations/charts';
+import { useChartContext } from '../ChartProvider';
 import { useTheme } from '@coinbase/cds-mobile/hooks/useTheme';
 
 import { DottedLine } from '../line';
