+++ packages/mobile-visualization/src/chart/bar/BarPlot.tsx
@@ -1,8 +1,10 @@
 import React, { memo, useMemo, useRef } from 'react';
 import { ClipPath, Defs, G, Rect } from 'react-native-svg';
 import type { ThemeVars } from '@coinbase/cds-common';
-import { defaultAxisId, useChartContext } from '@coinbase/cds-common/visualizations/charts';
+import { defaultAxisId } from '@coinbase/cds-common/visualizations/charts';
 import { generateRandomId } from '@coinbase/cds-utils';
+
+import { useChartContext } from '../ChartProvider';
 
 import type { BarComponent, BarProps } from './Bar';
 import type { BarSeries } from './BarChart';
