+++ packages/mobile-visualization/src/chart/line/GradientLine.tsx
@@ -1,7 +1,7 @@
 import { memo, useRef } from 'react';
 import { Defs, LinearGradient, Stop } from 'react-native-svg';
 import type { SharedProps } from '@coinbase/cds-common/types';
-import { useChartContext } from '@coinbase/cds-common/visualizations/charts';
+import { useChartContext } from '../ChartProvider';
 import { useTheme } from '@coinbase/cds-mobile/hooks/useTheme';
 import { generateRandomId } from '@coinbase/cds-utils';
