+++ packages/mobile-visualization/src/chart/line/__stories__/ReferenceLine.stories.tsx
@@ -3,7 +3,7 @@ import type { View } from 'react-native';
 import { Circle, type CircleProps, G, Path, Rect } from 'react-native-svg';
 import { assets } from '@coinbase/cds-common/internal/data/assets';
 import { sparklineInteractiveData } from '@coinbase/cds-common/internal/visualizations/SparklineInteractiveData';
-import { useChartContext } from '@coinbase/cds-common/visualizations/charts';
+import { useChartContext } from '../../ChartProvider';
 import { useTheme } from '@coinbase/cds-mobile';
 import { Example, ExampleScreen } from '@coinbase/cds-mobile/examples/ExampleScreen';
