+++ packages/web-visualization/src/chart/__stories__/TextStories.stories.tsx
@@ -1,9 +1,10 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import type { Rect } from '@coinbase/cds-common/types';
-import { isCategoricalScale, useChartContext } from '@coinbase/cds-common/visualizations/charts';
+import { isCategoricalScale } from '@coinbase/cds-common/visualizations/charts';
 import { Switch } from '@coinbase/cds-web/controls';
 
 import { XAxis, YAxis } from '../axis';
+import { useChartContext } from '../ChartProvider';
 import { LineChart } from '../line/LineChart';
 import { ChartText, SmartChartTextGroup, type TextLabelData } from '../text';
