+++ packages/web-visualization/src/chart/area/DottedArea.tsx
@@ -1,7 +1,7 @@
 import React, { memo, useRef } from 'react';
-import { useChartContext } from '@coinbase/cds-common/visualizations/charts';
 import { generateRandomId } from '@coinbase/cds-utils';
 
+import { useChartContext } from '../ChartProvider';
 import { Path, type PathProps } from '../Path';
 
 import type { AreaComponentProps } from './Area';
