+++ packages/web-visualization/src/chart/bar/BarStackGroup.tsx
@@ -1,8 +1,9 @@
 import React, { memo, useMemo } from 'react';
 import type { ThemeVars } from '@coinbase/cds-common';
-import { defaultAxisId, useChartContext } from '@coinbase/cds-common/visualizations/charts';
 import { getBarSizeAdjustment } from '@coinbase/cds-common/visualizations/charts/bar';
 import { isCategoricalScale } from '@coinbase/cds-common/visualizations/charts/scale';
+
+import { useChartContext } from '../ChartProvider';
 
 import type { BarComponent, BarProps } from './Bar';
 import type { BarSeries } from './BarChart';
