+++ packages/web-visualization/src/chart/bar/__stories__/BarChart.stories.tsx
@@ -1,15 +1,12 @@
 import React, { memo, useId } from 'react';
 import { candles as btcCandles } from '@coinbase/cds-common/internal/data/candles';
-import {
-  isCategoricalScale,
-  ScrubberContext,
-  useChartContext,
-} from '@coinbase/cds-common/visualizations/charts';
+import { isCategoricalScale, ScrubberContext } from '@coinbase/cds-common/visualizations/charts';
 import { VStack } from '@coinbase/cds-web/layout';
 import { Text } from '@coinbase/cds-web/typography';
 
 import { Chart } from '../..';
 import { XAxis, YAxis } from '../../axis';
+import { useChartContext } from '../../ChartProvider';
 import { ReferenceLine, SolidLine, type SolidLineProps } from '../../line';
 import { PeriodSelector } from '../../PeriodSelector';
 import { Scrubber } from '../../scrubber';
