+++ packages/web-visualization/src/chart/scrubber/ScrubberProvider.tsx
@@ -1,0 +1,242 @@
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import {
+  isCategoricalScale,
+  ScrubberContext,
+  type ScrubberContextValue,
+} from '@coinbase/cds-common/visualizations/charts';
+
+import { useChartContext } from '../ChartProvider';
+
+export type ScrubberProviderProps = {
+  children: React.ReactNode;
+  /**
+   * A reference to the root SVG element, where interaction event handlers will be attached.
+   */
+  svgRef: React.RefObject<SVGSVGElement> | null;
+  /**
+   * Enables scrubbing interactions (mouse and keyboard highlighting).
+   * When true, allows highlighting and makes scrubber components interactive.
+   * @default false
+   */
+  enableScrubbing?: boolean;
+  /**
+   * Callback fired when the highlighted item changes.
+   * Receives the dataIndex of the highlighted item or null when no item is highlighted.
+   */
+  onScrubberPosChange?: (dataIndex: number | null) => void;
+};
+
+/**
+ * A component which encapsulates the ScrubberContext.
+ * It depends on a ChartContext in order to provide accurate mouse tracking.
+ */
+export const ScrubberProvider: React.FC<ScrubberProviderProps> = ({
+  children,
+  svgRef,
+  enableScrubbing = false,
+  onScrubberPosChange,
+}) => {
+  const chartContext = useChartContext();
+
+  if (!chartContext) {
+    throw new Error('ScrubberProvider must be used within a ChartContext');
+  }
+
+  const { getXScale, getXAxis, series } = chartContext;
+  const [highlightedIndex, setHighlightedIndex] = useState<number | undefined>(undefined);
+
+  const getDataIndexFromX = useCallback(
+    (mouseX: number): number => {
+      const xScale = getXScale();
+      const xAxis = getXAxis();
+
+      if (!xScale || !xAxis) return 0;
+
+      if (isCategoricalScale(xScale)) {
+        const categories = xScale.domain?.() ?? xAxis.data ?? [];
+        const bandwidth = xScale.bandwidth?.() ?? 0;
+        let closestIndex = 0;
+        let closestDistance = Infinity;
+        for (let i = 0; i < categories.length; i++) {
+          const xPos = xScale(i);
+          if (xPos !== undefined) {
+            const distance = Math.abs(mouseX - (xPos + bandwidth / 2));
+            if (distance < closestDistance) {
+              closestDistance = distance;
+              closestIndex = i;
+            }
+          }
+        }
+        return closestIndex;
+      } else {
+        // For numeric scales with axis data, find the nearest data point
+        const axisData = xAxis.data;
+        if (axisData && Array.isArray(axisData) && typeof axisData[0] === 'number') {
+          // We have numeric axis data - find the closest data point
+          const numericData = axisData as number[];
+          let closestIndex = 0;
+          let closestDistance = Infinity;
+
+          for (let i = 0; i < numericData.length; i++) {
+            const xValue = numericData[i];
+            const xPos = xScale(xValue);
+            if (xPos !== undefined) {
+              const distance = Math.abs(mouseX - xPos);
+              if (distance < closestDistance) {
+                closestDistance = distance;
+                closestIndex = i;
+              }
+            }
+          }
+          return closestIndex;
+        } else {
+          const xValue = xScale.invert(mouseX);
+          const dataIndex = Math.round(xValue);
+          const domain = xAxis.domain;
+          return Math.max(domain.min ?? 0, Math.min(dataIndex, domain.max ?? 0));
+        }
+      }
+    },
+    [getXScale, getXAxis],
+  );
+
+  const handleMouseMove = useCallback(
+    (event: MouseEvent) => {
+      if (!enableScrubbing || !series || series.length === 0) return;
+
+      const target = event.currentTarget as SVGSVGElement;
+      const rect = target.getBoundingClientRect();
+      const mouseX = event.clientX - rect.left;
+
+      const dataIndex = getDataIndexFromX(mouseX);
+
+      if (dataIndex !== highlightedIndex) {
+        setHighlightedIndex(dataIndex);
+        onScrubberPosChange?.(dataIndex);
+      }
+    },
+    [enableScrubbing, series, getDataIndexFromX, highlightedIndex, onScrubberPosChange],
+  );
+
+  const handleMouseLeave = useCallback(() => {
+    if (!enableScrubbing) return;
+    setHighlightedIndex(undefined);
+    onScrubberPosChange?.(null);
+  }, [enableScrubbing, onScrubberPosChange]);
+
+  const handleKeyDown = useCallback(
+    (event: KeyboardEvent) => {
+      if (!enableScrubbing) return;
+
+      const xScale = getXScale();
+      const xAxis = getXAxis();
+
+      if (!xScale || !xAxis) return;
+
+      const isBand = isCategoricalScale(xScale);
+
+      // Determine the actual data indices we can navigate to
+      let minIndex: number;
+      let maxIndex: number;
+      let dataPoints: number | undefined;
+
+      if (isBand) {
+        // For categorical scales, use the categories
+        const categories = xScale.domain?.() ?? xAxis.data ?? [];
+        minIndex = 0;
+        maxIndex = Math.max(0, categories.length - 1);
+        dataPoints = categories.length;
+      } else {
+        // For numeric scales, check if we have specific data points
+        const axisData = xAxis.data;
+        if (axisData && Array.isArray(axisData)) {
+          // We have specific data points - use their indices
+          minIndex = 0;
+          maxIndex = Math.max(0, axisData.length - 1);
+          dataPoints = axisData.length;
+        } else {
+          // Fall back to domain-based navigation for continuous scales without specific data
+          const domain = xAxis.domain;
+          minIndex = domain.min ?? 0;
+          maxIndex = domain.max ?? 0;
+          dataPoints = maxIndex - minIndex + 1;
+        }
+      }
+
+      const currentIndex = highlightedIndex ?? minIndex;
+      const dataRange = maxIndex - minIndex;
+
+      // Multi-step jump when shift is held (10% of data range, minimum 1, maximum 10)
+      const multiSkip = event.shiftKey;
+      const stepSize = multiSkip ? Math.min(10, Math.max(1, Math.floor(dataRange * 0.1))) : 1;
+
+      let newIndex: number | undefined;
+
+      switch (event.key) {
+        case 'ArrowLeft':
+          event.preventDefault();
+          newIndex = Math.max(minIndex, currentIndex - stepSize);
+          break;
+        case 'ArrowRight':
+          event.preventDefault();
+          newIndex = Math.min(maxIndex, currentIndex + stepSize);
+          break;
+        case 'Home':
+          event.preventDefault();
+          newIndex = minIndex;
+          break;
+        case 'End':
+          event.preventDefault();
+          newIndex = maxIndex;
+          break;
+        case 'Escape':
+          event.preventDefault();
+          newIndex = undefined; // Clear highlighting
+          break;
+        default:
+          return; // Don't handle other keys
+      }
+
+      if (newIndex !== highlightedIndex) {
+        setHighlightedIndex(newIndex);
+        onScrubberPosChange?.(newIndex ?? null);
+      }
+    },
+    [enableScrubbing, getXScale, getXAxis, highlightedIndex, onScrubberPosChange],
+  );
+
+  // Attach event listeners to SVG element
+  useEffect(() => {
+    if (!svgRef?.current || !enableScrubbing) return;
+
+    const svg = svgRef.current;
+
+    // Add event listeners
+    svg.addEventListener('mousemove', handleMouseMove);
+    svg.addEventListener('mouseleave', handleMouseLeave);
+    svg.addEventListener('keydown', handleKeyDown);
+
+    // Add tabIndex for keyboard support
+    svg.setAttribute('tabIndex', '0');
+
+    return () => {
+      svg.removeEventListener('mousemove', handleMouseMove);
+      svg.removeEventListener('mouseleave', handleMouseLeave);
+      svg.removeEventListener('keydown', handleKeyDown);
+      if (svg.hasAttribute('tabIndex')) {
+        svg.removeAttribute('tabIndex');
+      }
+    };
+  }, [svgRef, enableScrubbing, handleMouseMove, handleMouseLeave, handleKeyDown]);
+
+  const contextValue: ScrubberContextValue = useMemo(
+    () => ({
+      scrubbingEnabled: enableScrubbing,
+      highlightedIndex,
+      updateHighlightedIndex: setHighlightedIndex,
+    }),
+    [enableScrubbing, highlightedIndex],
+  );
+
+  return <ScrubberContext.Provider value={contextValue}>{children}</ScrubberContext.Provider>;
+};
